Bill of Rights as submitted to the States
Congress of the United States begun and held at the City of New-York, on
Wednesday the fourth of March, one thousand seven hundred and eighty nine.

The Conventions of a number of the States, having at the time of their
adopting the Constitution, expressed a desire, in order to prevent
misconstruction or abuse of its powers, that further declaratory and
restrictive clauses should be added:  And as extending the ground of public
confidence in the Government, will best ensure the beneficent ends of its
institution.

Resolved by the Senate and House of Representatives of the United States of
America, in Congress assembled, two thirds of both Houses concurring, that
the following Articles be proposed to the Legislatures of the several
States as amendments to the Constitution of the United States, all, or any
of which articles, when ratified by three fourths of the said Legislatures,
to be valid to all intents and purposes, as part of the said Constitution.

Frederick Augustus Muhlenberg, Speaker of the House
John Adams, Vice-President of the United States and President of the Senate

Attest John Beckley, Clerk of the House of Representatives
Sam. A. Otis Secretary of the Senate