Constitution as submitted to Congress and the States
In Convention Monday September 17ᵗʰ 1787.

Present the States of New Hampshire, Massachusetts, Connecticut,
	Mr. Hamilton from New York, New Jersey, Pennsylvania, Delaware,
	Maryland, Virginia, North Carolina, South Carolina and Georgia.

Resolved,	That the proceeding Constitution be laid before the
United States in Congress assembled, and that it is the Opinion of this
Convention, that it should afterwards be submitted to a Convention of
Delegates, chosen in each State by the People thereof, under the
Recommendation of its Legislature, for their Assent and Ratification; and
that each Convention assenting to, and ratifying the Same, should give
Notice thereof to the United States in Congress assembled.

Resolved,	That it is the Opinion of this Convention, that as soon as
the Conventions of nine States shall have ratified this Constitution, the
United States in Congress assembled should fix a Day on which Electors
should be appointed by the States which shall have ratified the same, and a
Day on which the Electors should assemble to vote for the President, and
the Time and Place for commencing Proceedings under this Constitution.
That after such Publication the Electors should be appointed, and the
Senators and Representatives elected: That the Electors should meet on the
Day fixed for the Election of the President, and should transmit their
Votes certified, signed, sealed and directed, as the Constitution requires,
to the Secretary of the United States in Congress assembled, that the
Senators and Representatives should convene at the Time and Place assigned;
that the Senators should appoint a President of the Senate, for the sole
Purpose of receiving, opening and counting the Votes for President; and,
that after he shall be chosen, the Congress, together with the President,
should, without Delay, proceed to execute this Constitution.

By the Unanimous Order of the Convention,
Gᵒ Washington, Presidᵗ

W. Jackson Secretary.